+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Added
+
+- Silence errors with `mypy`.
+
 ## 1.4.2 (2024-07-01)
 
 ### Fixed